Define Your Research Design and Create Code Directory with Step Templates

This generates a directory at rel_dir (defaulting to "./code"). In this directory it will create a set of files with templates for all code steps or a joint file containing all step templates.

define_design(steps, rel_dir = "./code", one_file = FALSE,
  one_file_name = "design_steps.R")

Arguments

steps
A character vector containing the names for the step functions. All members need to be valid R names.

rel_dir
A relative path to a directory in which you want to create the step templates.

one_file
If TRUE, then all templates will be stored in a file with the name one_file_name. If FALSE (the default), each step template will be stored in a separate file with the step name acting as file name.

one_file_name
The name of the code file conating the step templates if all step templates are stored in one file. Defaults to "design_steps.R".

Value

The steps parameter.

Details

See the vignette of the package for further details.

Examples

# NOT RUN {
  print("Sorry. No examples yet.")
# }