Plot Researcher Degrees of Freedom Estimates' Density

Visualizes the estimates generated by exhausting all documented researchers degrees of freedom.

plot_rdf_estimate_density(df, est, lb, ub, color = "blue", n = 30)

Arguments

df
The data frame as generated by exhaust_design().

est
A character value to indicate the variable in df that contains the pont estimate that you want to plot.

lb
A character value to indicate the variable in df that contains the lower bound of your estimate.

ub
A character value to indicate the variable in df that contains the upper bound of your estimate.

color
Which color do you want your estimate density to have?

n
The condidence interval is currently naively estimate by uniformly distributing n observations between lb and ub

Value

A ggplot object containing the plot.

Details

Currently naively assumes a uniform distribution between lower and upper bound of the estimate to display the confidence interval. See the vignette of the package for further details on how to implement the RDF workflow.

Examples

# NOT RUN {
  print("Sorry. No examples yet.")
# }