Plot Estimates by Choice

Visualizes the variance of estimates conditional on a selected discrete and continous choice.

Usage

plot_rdf_estimates_by_choice(
  df,
  est,
  lb,
  ub,
  dchoice,
  color = NULL,
  order = NULL,
  width = 1
)

Arguments

df

The data frame as generated by exhaust_design().

est

A character value to indicate the variable in df that contains the pont estimate that you want to plot.

lb

A character value indicating the variable in df that contains the lower bound of the estimate.

ub

A character value indicating the variable in df that contains the upper bound of the estimate.

dchoice

A character value to indicate the discrete choice that you want to partition your estimates on.

color

A character value to indicate a choice for the coloring of the point ranges. Can be continous or discrete.

order

A character value to indicate a choice for ordering the point ranges. If NULL (the default), the point ranges will be ordered by estimate magnitude.

width

The width parameter for position_dodge2 to spread out the point ranges within the discrete choice area.

Value

A ggplot object containing the plot.

Details

See the vignette of the package for further details on how to implement the RDF workflow.

Examples

if (FALSE) {
  print("Sorry. No examples yet.")
}