Prepare A Flow Chart of Your Research Design

Generates a DiagrammeR flow chart of your research design. Output can be saved to a file or directly viewed.

prepare_design_flow_chart(d, file_name = NULL, landscape = FALSE,
  color_step = "red", color_dchoice = "green",
  color_cchoice = "blue", color_result = "orange")

Arguments

d
A character vector of the research design steps function names

file_name
A file name for an output PDF file relative to the working directory.

landscape
Do you want the flow chart to be generated horizontally instead of vertically. Defaults to FALSE.

color_step
The fill color to indicate a design step.

color_dchoice
The fill color to indicate a discrete choice.

color_cchoice
The fill color to indicate a continuous choice.

color_result
The fill color to indicate a result.

Value

the rendered diagram or, if file_name is specified, the return code from DiagrammeRsvg::export_graph()

Examples

# NOT RUN {
  print("Sorry. No examples yet.")
# }