An Interactive Specification Curve

A shiny based web app that allows you to explore your researcher degrees of freedom's specification curve interactively.

shiny_rdf_spec_curve(ests, spec_curve_parms, design = NULL,
  rel_dir = NULL, start_input = NULL, libs = NULL,
  regression_cutoff = 5, default_choices = NULL,
  title = "A Shiny Specification Curve", abstract = NULL)

Arguments

ests
The data frame provided by exhaust_design().

spec_curve_parms
A list containing additional paramters that will be passed on to plot_rdf_spec_curve().

design
if not NULL it takes the design that was used to generate the estimates. In this case, you also need to specify the rel_dir and start_input parameter below. The shiny app will then display full regresssion results when you select choices that generate less than regression_cutoff estimates.

rel_dir
The path to the code directory. See above.

start_input
The parameters that you pass to the first design step. See above.

libs
A vector containing additional packages that need to be attached to run the design. NOTE: While this works fine when you host the shiny app yourself, shinyapps.io fails to include the listed packages when deploying the app. So, if you plan to host your app on shinyapps.io, you are better served including library() calls in your design code files or to use the :: operator in your code.

regression_cutoff
If your choices generate less or equal estimates, the display will switch to normal regression output (needs parameters above to be not NULL).

default_choices
A list containing choices that you want the app to start with. If NULL, it will start with all choices included.

title
The title of the shiny app.

abstract
Text that will be displayed by the app. Wrapped into HTML() so that you can use HTML code.

Examples

# NOT RUN {
  print("Sorry. No examples yet.")
# }